+++ MapboxNavigation/ImageDownloader.swift
@@ -4,26 +4,29 @@ typealias ImageDownloadCompletionBlock = (UIImage?, Data?, Error?) -> Void
 
 protocol ReentrantImageDownloader {
     func downloadImage(with url: URL, completion: ImageDownloadCompletionBlock?) -> Void
-    func activeOperationWithURL(_ url: URL) -> ImageDownload?
+    func activeOperation(with url: URL) -> ImageDownload?
     func setOperationType(_ operationType: ImageDownload.Type?)
 }
 
 class ImageDownloader: NSObject, ReentrantImageDownloader, URLSessionDataDelegate {
     private var sessionConfiguration: URLSessionConfiguration = URLSessionConfiguration.default
-    private var operationType: ImageDownload.Type = ImageDownloadOperation.self
 
-    private var queue: OperationQueue
     lazy private var urlSession: URLSession = {
         return URLSession.init(configuration: sessionConfiguration, delegate: self, delegateQueue: nil)
     }()
 
-    private var headers: [String: String] = ["Accept": "image/*;q=0.8"]
+    private var downloadQueue: OperationQueue
+    private var accessQueue: DispatchQueue
 
+    private var operationType: ImageDownload.Type = ImageDownloadOperation.self
     private var operations: [URL: ImageDownload] = [:]
 
+    private var headers: [String: String] = ["Accept": "image/*;q=0.8"]
+
     override init() {
-        self.queue = OperationQueue()
-        self.queue.name = Bundle.mapboxNavigation.bundleIdentifier! + ".ImageDownloader"
+        self.downloadQueue = OperationQueue()
+        self.downloadQueue.name = Bundle.mapboxNavigation.bundleIdentifier! + ".ImageDownloader"
+        self.accessQueue = DispatchQueue(label: Bundle.mapboxNavigation.bundleIdentifier! + ".ImageDownloaderInternal")
     }
 
     convenience init(sessionConfiguration: URLSessionConfiguration? = nil, operationType: ImageDownload.Type? = nil) {
@@ -38,16 +41,22 @@ class ImageDownloader: NSObject, ReentrantImageDownloader, URLSessionDataDelegat
         }
     }
 
+    deinit {
+        self.downloadQueue.cancelAllOperations()
+    }
+
     func downloadImage(with url: URL, completion: ImageDownloadCompletionBlock?) {
-        let request: URLRequest = urlRequestWithURL(url)
+        let request: URLRequest = urlRequest(with: url)
         var operation: ImageDownload
-        if let activeOperation = activeOperationWithURL(url) {
+        if let activeOperation = activeOperation(with: url) {
             operation = activeOperation
         } else {
             operation = operationType.init(request: request, in: self.urlSession)
-            self.operations[url] = operation
+            accessQueue.async(flags: .barrier) {
+                self.operations[url] = operation
+            }
             if let operation = operation as? Operation {
-                self.queue.addOperation(operation)
+                self.downloadQueue.addOperation(operation)
             }
         }
         if let completion = completion {
@@ -55,14 +64,19 @@ class ImageDownloader: NSObject, ReentrantImageDownloader, URLSessionDataDelegat
         }
     }
 
-    func activeOperationWithURL(_ url: URL) -> ImageDownload? {
-        if let operation = operations[url], !operation.isFinished {
-            return operation
+    func activeOperation(with url: URL) -> ImageDownload? {
+        var activeOperation: ImageDownload?
+
+        accessQueue.sync {
+            if let operation = operations[url], !operation.isFinished {
+                activeOperation = operation
+            }
         }
-        return nil
+
+        return activeOperation
     }
 
-    private func urlRequestWithURL(_ url: URL) -> URLRequest {
+    private func urlRequest(with url: URL) -> URLRequest {
         var request = URLRequest(url: url)
         request.allHTTPHeaderFields = self.headers
         request.cachePolicy = .reloadIgnoringCacheData
@@ -80,7 +94,7 @@ class ImageDownloader: NSObject, ReentrantImageDownloader, URLSessionDataDelegat
     // MARK: URLSessionDataDelegate
 
     func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didReceive response: URLResponse, completionHandler: @escaping (URLSession.ResponseDisposition) -> Void) {
-        guard let response: HTTPURLResponse = response as? HTTPURLResponse, let url = response.url, let operation: ImageDownload = operations[url] else {
+        guard let response: HTTPURLResponse = response as? HTTPURLResponse, let url = response.url, let operation: ImageDownload = activeOperation(with: url) else {
             completionHandler(.cancel)
             return
         }
@@ -88,18 +102,20 @@ class ImageDownloader: NSObject, ReentrantImageDownloader, URLSessionDataDelegat
     }
 
     func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didReceive data: Data) {
-        guard let url = dataTask.originalRequest?.url, let operation: ImageDownload = operations[url] else {
+        guard let url = dataTask.originalRequest?.url, let operation: ImageDownload = activeOperation(with: url) else {
             return
         }
         operation.urlSession?(session, dataTask: dataTask, didReceive: data)
     }
 
     func urlSession(_ session: URLSession, task: URLSessionTask, didCompleteWithError error: Error?) {
-        guard let url = task.originalRequest?.url, let operation: ImageDownload = operations[url] else {
+        guard let url = task.originalRequest?.url, let operation: ImageDownload = activeOperation(with: url) else {
             return
         }
         operation.urlSession?(session, task: task, didCompleteWithError: error)
-        operations[url] = nil
+        accessQueue.async {
+            self.operations[url] = nil
+        }
     }
 
 }